Extract MultiParentCasperDeploySpec
