+++ casper/src/test/scala/coop/rchain/casper/HashSetCasperTest.scala
@@ -70,89 +70,7 @@ class HashSetCasperTest extends FlatSpec with Matchers with Inspectors {
 
   //put a new casper instance at the start of each
   //test since we cannot reset it
-  "HashSetCasper" should "accept deploys" in effectTest {
-    val node = HashSetCasperTestNode.standaloneEff(genesis, validatorKeys.head)
-    import node._
-    implicit val timeEff = new LogicalTime[Effect]
-
-    for {
-      deploy <- ConstructDeploy.basicDeployData[Effect](0)
-      _      <- MultiParentCasper[Effect].deploy(deploy)
-
-      _      = logEff.infos.size should be(2)
-      result = logEff.infos(1).contains("Received Deploy") should be(true)
-      _      <- node.tearDown()
-    } yield result
-  }
-
-  it should "not allow deploy if deploy is missing signature" in effectTest {
-    val node             = HashSetCasperTestNode.standaloneEff(genesis, validatorKeys.head)
-    val casper           = node.casperEff
-    implicit val timeEff = new LogicalTime[Effect]
-
-    for {
-      correctDeploy   <- ConstructDeploy.basicDeployData[Effect](0)
-      incorrectDeploy = correctDeploy.withSig(ByteString.EMPTY)
-      deployResult    <- casper.deploy(incorrectDeploy)
-      _               <- node.tearDown()
-    } yield deployResult should be(Left(MissingSignature))
-  }
-
-  it should "not allow deploy if deploy is missing signature algorithm" in effectTest {
-    val node             = HashSetCasperTestNode.standaloneEff(genesis, validatorKeys.head)
-    val casper           = node.casperEff
-    implicit val timeEff = new LogicalTime[Effect]
-
-    for {
-      correctDeploy   <- ConstructDeploy.basicDeployData[Effect](0)
-      incorrectDeploy = correctDeploy.withSigAlgorithm("")
-      deployResult    <- casper.deploy(incorrectDeploy)
-      _               <- node.tearDown()
-    } yield deployResult should be(Left(MissingSignatureAlgorithm))
-  }
-
-  it should "not allow deploy if deploy is missing user" in effectTest {
-    val node             = HashSetCasperTestNode.standaloneEff(genesis, validatorKeys.head)
-    val casper           = node.casperEff
-    implicit val timeEff = new LogicalTime[Effect]
-
-    for {
-      correctDeploy   <- ConstructDeploy.basicDeployData[Effect](0)
-      incorrectDeploy = correctDeploy.withDeployer(ByteString.EMPTY)
-      deployResult    <- casper.deploy(incorrectDeploy)
-      _               <- node.tearDown()
-    } yield deployResult should be(Left(MissingUser))
-  }
-
-  it should "not allow deploy if deploy is holding non-existing algorithm" in effectTest {
-    val node             = HashSetCasperTestNode.standaloneEff(genesis, validatorKeys.head)
-    val casper           = node.casperEff
-    implicit val timeEff = new LogicalTime[Effect]
-
-    for {
-      correctDeploy   <- ConstructDeploy.basicDeployData[Effect](0)
-      incorrectDeploy = correctDeploy.withSigAlgorithm("SOME_RANDOME_STUFF")
-      deployResult    <- casper.deploy(incorrectDeploy)
-      _               <- node.tearDown()
-    } yield deployResult should be(Left(UnknownSignatureAlgorithm("SOME_RANDOME_STUFF")))
-  }
-
-  it should "not allow deploy if deploy is incorrectly signed" in effectTest {
-    val node             = HashSetCasperTestNode.standaloneEff(genesis, validatorKeys.head)
-    val casper           = node.casperEff
-    implicit val timeEff = new LogicalTime[Effect]
-
-    for {
-      correctDeploy <- ConstructDeploy.basicDeployData[Effect](0)
-      incorrectDeploy = correctDeploy.withSig(
-        ByteString.copyFrom(correctDeploy.sig.toByteArray.reverse)
-      )
-      deployResult <- casper.deploy(incorrectDeploy)
-      _            <- node.tearDown()
-    } yield deployResult should be(Left(SignatureVerificationFailed))
-  }
-
-  it should "not allow multiple threads to process the same block" in {
+  "MultiParentCasper" should "not allow multiple threads to process the same block" in {
     val scheduler = Scheduler.fixedPool("three-threads", 3)
     val node      = HashSetCasperTestNode.standaloneEff(genesis, validatorKeys.head)(scheduler)
     val casper    = node.casperEff
